Gave up and went with basic option for dark mode (#2)
Had to settle for a third party dark theme as Bootstrap amazingly does not have native dark theme support in almost-2022. This one is not great and has bad contrast in some areas, but it is one of the only ones that seems actively maintained and is available on npm (don't want to use CDN).

Could not implement a manual toggle (so for now, the dark theme automatically follows the system setting only), because importing the dark theme file using `@import` doesn't work right from inside a class; the styles don't get applied correctly even when put on the root `html` element (or `body` or app-component). So the `@import` can only be placed in the root of `styles.scss` - meaning it automatically follows the system setting only and cannot be restricted/forced using a classname.

As for the aforementioned low contrast, it is not possible to override some theme classes with other bootstrap classes, because for now, SCSS does not support the `@extend` operator inside of a media query. So the only solution would be to re-define all relevant bootstrap styles again manually, which is ridiculous and defeats the purpose of relying on these libraries for styling.

TL;DR: Dark mode technically works but is ugly and automatic instead of vibrant and with a toggle; needs future attention.